feat: persist store implementation
