+++ package.json
@@ -14,6 +14,8 @@
     "react-scripts": "5.0.1",
     "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
+    "redux-persist": "^6.0.0",
+    "redux-thunk": "^3.1.0",
     "reselect": "^5.1.0",
     "sass": "^1.70.0",
     "styled-components": "^6.1.8",
