+++ src/utils/firebase/firebase.util.js
@@ -70,13 +70,6 @@ export const getCategoriesAndDocuments = async () => {
   const querySnapshot = await getDocs(q);
 
   return querySnapshot.docs.map((docRef) => docRef.data());
-  // const categoryMap = querySnapshot.docs.reduce((acc, docSnapShot) => {
-  //   const { title, items } = docSnapShot.data();
-  //   acc[title.toLowerCase()] = items;
-  //   return acc;
-  // }, {});
-
-  // return categoryMap;
 };
 
 export const createUserDocumentFromAuth = async (userAuth) => {
